requirements.txt
Flask>=3.0,<4.0
gunicorn>=22.0,<24.0
psycopg[binary]>=3.2,<4.0
google-auth>=2.30,<3.0
PyJWT>=2.8,<3.0
requests>=2.31,<3.0
